+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Pashmak Changelog
 
+## next release
+
+- fixed `$__file__` and `$__dir__` variables bug in windows
+
 ## 0.5.2 (2020-12-5)
 
 - fixed namespace bug cannot using variables from another namespace inside a namespace